+++ packages/wallet-management/src/walletIcons/index.ts
@@ -26,6 +26,9 @@ import trust from './trust.svg';
 import walletConnect from './walletConnect.svg';
 import walletio from './walletio.svg';
 import xdefi from './xdefi.svg';
+import bitGet from './bitget.svg';
+import okx from './okx.svg';
+import gate from './gate.svg';
 
 export const walletIcons = {
   mathwallet,
@@ -33,6 +36,8 @@ export const walletIcons = {
   alphawallet,
   frontier,
   blockwallet,
+  bitGet,
+  gate,
   binance,
   placeholder,
   bitpie,
@@ -44,6 +49,7 @@ export const walletIcons = {
   imtoken,
   liquality,
   metamask,
+  okx,
   oneInch,
   opera,
   ownbit,